requirements.txt
fastapi==0.111.0
langchain==0.1.17
langchain_community==0.0.37
langchain_core==0.1.52
langchain_openai==0.1.6
openai==1.26.0
pydantic==2.7.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
SQLAlchemy==2.0.29
streamlit==1.33.0
todoist_api_python==2.1.3
